Add combination calculator and README instructions
This commit adds a new Go module for calculating the number of combinations of alphanumeric characters which can be formed for a given length. The main execution file, `main.go`, takes an input from the user for the length of the set and calculates the number of combinations. The README contains detailed instructions on how to run the program and examples of the results for inputs between 1 and 10. The `.gitignore` file has been updated to exclude irrelevant files from the project repository.

This new functionality will provide users with a quick way to understand the potential complexity of a chosen alphanumeric character set. The README serves as a user-friendly guide to utilizing this new combination calculation feature.